mgs_author cast as int sothat does not return all messages for empty id

+++ admin/class-fep-admin-pages.php
@@ -372,7 +372,7 @@ class Fep_Admin_Pages {
 			'paged'      => (int) $page,
 			'per_page'   => 100,
 			'mgs_status' => 'any',
-			'mgs_author' => fep_get_userdata( $email_address, 'ID', 'email' ),
+			'mgs_author' => (int) fep_get_userdata( $email_address, 'ID', 'email' ),
 		);
 		$messages     = fep_get_messages( $args );
 		$export_items = array();
